Big refactoring to update most of the functions in fitting.py.
Still todo
1) Multi-fit rewrites
