+++ submm/KIDs/res/fitting.py
@@ -23,146 +23,17 @@ import inspect
 
 import numpy as np
 import scipy.optimize as optimization
-import matplotlib.pyplot as plt
 
 from submm.KIDs.res.fit_funcs import linear_mag, nonlinear_mag, nonlinear_iq, nonlinear_iq_for_fitter, \
-    NonlinearIQRes, Fit, ResSet
-from submm.KIDs.res.utils import print_fit_string_nonlinear_iq, print_fit_string_nonlinear_mag, \
-    print_fit_string_linear_mag, amplitude_normalization, guess_x0_iq_nonlinear, guess_x0_mag_nonlinear, \
-    guess_x0_iq_nonlinear_sep, guess_x0_mag_nonlinear_sep, calc_qc_qi
-
-
-def brute_force_linear_mag_fit(f_hz, z, ranges, n_grid_points, error=None, plot=False):
-    """
-    Parameters
-    ----------
-    f_hz : numpy.array
-        frequencies Hz
-    z : numpy.array
-        complex or abs of s21
-    ranges : numpy.array
-        The ranges for each parameter as in:
-        np.asarray(([f_low,Qr_low,amp_low,phi_low,b0_low],[f_high,Qr_high,amp_high,phi_high,b0_high]))
-    n_grid_points: int
-        How finely to sample each parameter space. This can be very slow for n>10, an increase by a factor of 2 will
-        take 2**5 times longer to marginalize over you must minimize over the unwanted axes of sum_dev
-        i.e. for fr np.min(np.min(np.min(np.min(fit['sum_dev'],axis = 4),axis = 3),axis = 2),axis = 1)
-    error : numpy.array, optional (default = None)
-        The error on the complex or abs of s21, used for weighting squares
-    plot : bool, optional (default = False)
-        If true, will plot the fit and the data
-    """
-    if error is None:
-        error = np.ones(len(f_hz))
-
-    fs = np.linspace(ranges[0][0], ranges[1][0], n_grid_points)
-    Qrs = np.linspace(ranges[0][1], ranges[1][1], n_grid_points)
-    amps = np.linspace(ranges[0][2], ranges[1][2], n_grid_points)
-    phis = np.linspace(ranges[0][3], ranges[1][3], n_grid_points)
-    b0s = np.linspace(ranges[0][4], ranges[1][4], n_grid_points)
-    evaluated_ranges = np.vstack((fs, Qrs, amps, phis, b0s))
-
-    a, b, c, d, e = np.meshgrid(fs, Qrs, amps, phis, b0s, indexing="ij")  # always index ij
-
-    evaluated = linear_mag(f_hz, a, b, c, d, e)
-    data_values = np.reshape(np.abs(z) ** 2, (abs(z).shape[0], 1, 1, 1, 1, 1))
-    error = np.reshape(error, (abs(z).shape[0], 1, 1, 1, 1, 1))
-    sum_dev = np.sum(((np.sqrt(evaluated) - np.sqrt(data_values)) ** 2 / error ** 2),
-                     axis=0)  # comparing in magnitude space rather than magnitude squared
-
-    min_index = np.where(sum_dev == np.min(sum_dev))
-    index1 = min_index[0][0]
-    index2 = min_index[1][0]
-    index3 = min_index[2][0]
-    index4 = min_index[3][0]
-    index5 = min_index[4][0]
-    fit_values = np.asarray((fs[index1], Qrs[index2], amps[index3], phis[index4], b0s[index5]))
-    fit_values_names = ('f0', 'Qr', 'amp', 'phi', 'b0')
-    fit_result = linear_mag(f_hz, fs[index1], Qrs[index2], amps[index3], phis[index4], b0s[index5])
-
-    marginalized_1d = np.zeros((5, n_grid_points))
-    marginalized_1d[0, :] = np.min(np.min(np.min(np.min(sum_dev, axis=4), axis=3), axis=2), axis=1)
-    marginalized_1d[1, :] = np.min(np.min(np.min(np.min(sum_dev, axis=4), axis=3), axis=2), axis=0)
-    marginalized_1d[2, :] = np.min(np.min(np.min(np.min(sum_dev, axis=4), axis=3), axis=1), axis=0)
-    marginalized_1d[3, :] = np.min(np.min(np.min(np.min(sum_dev, axis=4), axis=2), axis=1), axis=0)
-    marginalized_1d[4, :] = np.min(np.min(np.min(np.min(sum_dev, axis=3), axis=2), axis=1), axis=0)
-
-    marginalized_2d = np.zeros((5, 5, n_grid_points, n_grid_points))
-    # 0 _
-    # 1 x _
-    # 2 x x _
-    # 3 x x x _ 
-    # 4 x x x x _
-    #  0 1 2 3 4
-    marginalized_2d[0, 1, :] = marginalized_2d[1, 0, :] = np.min(np.min(np.min(sum_dev, axis=4), axis=3), axis=2)
-    marginalized_2d[2, 0, :] = marginalized_2d[0, 2, :] = np.min(np.min(np.min(sum_dev, axis=4), axis=3), axis=1)
-    marginalized_2d[2, 1, :] = marginalized_2d[1, 2, :] = np.min(np.min(np.min(sum_dev, axis=4), axis=3), axis=0)
-    marginalized_2d[3, 0, :] = marginalized_2d[0, 3, :] = np.min(np.min(np.min(sum_dev, axis=4), axis=2), axis=1)
-    marginalized_2d[3, 1, :] = marginalized_2d[1, 3, :] = np.min(np.min(np.min(sum_dev, axis=4), axis=2), axis=0)
-    marginalized_2d[3, 2, :] = marginalized_2d[2, 3, :] = np.min(np.min(np.min(sum_dev, axis=4), axis=1), axis=0)
-    marginalized_2d[4, 0, :] = marginalized_2d[0, 4, :] = np.min(np.min(np.min(sum_dev, axis=3), axis=2), axis=1)
-    marginalized_2d[4, 1, :] = marginalized_2d[1, 4, :] = np.min(np.min(np.min(sum_dev, axis=3), axis=2), axis=0)
-    marginalized_2d[4, 2, :] = marginalized_2d[2, 4, :] = np.min(np.min(np.min(sum_dev, axis=3), axis=1), axis=0)
-    marginalized_2d[4, 3, :] = marginalized_2d[3, 4, :] = np.min(np.min(np.min(sum_dev, axis=2), axis=1), axis=0)
-
-    if plot:
-        levels = [2.3, 4.61]  # delta chi squared two parameters 68 90 % confidence
-        fig_fit = plt.figure(-1)
-        axs = fig_fit.subplots(5, 5)
-        for i in range(0, 5):  # y starting from top
-            for j in range(0, 5):  # x starting from left
-                if i > j:
-                    # plt.subplot(5,5,i+1+5*j)
-                    # axs[i, j].set_aspect('equal', 'box')
-                    extent = [evaluated_ranges[j, 0], evaluated_ranges[j, n_grid_points - 1], evaluated_ranges[i, 0],
-                              evaluated_ranges[i, n_grid_points - 1]]
-                    axs[i, j].imshow(marginalized_2d[i, j, :] - np.min(sum_dev), extent=extent, origin='lower',
-                                     cmap='jet')
-                    axs[i, j].contour(evaluated_ranges[j], evaluated_ranges[i],
-                                      marginalized_2d[i, j, :] - np.min(sum_dev), levels=levels, colors='white')
-                    axs[i, j].set_ylim(evaluated_ranges[i, 0], evaluated_ranges[i, n_grid_points - 1])
-                    axs[i, j].set_xlim(evaluated_ranges[j, 0], evaluated_ranges[j, n_grid_points - 1])
-                    axs[i, j].set_aspect((evaluated_ranges[j, 0] - evaluated_ranges[j, n_grid_points - 1]) / (
-                            evaluated_ranges[i, 0] - evaluated_ranges[i, n_grid_points - 1]))
-                    if j == 0:
-                        axs[i, j].set_ylabel(fit_values_names[i])
-                    if i == 4:
-                        axs[i, j].set_xlabel("\n" + fit_values_names[j])
-                    if i < 4:
-                        axs[i, j].get_xaxis().set_ticks([])
-                    if j > 0:
-                        axs[i, j].get_yaxis().set_ticks([])
-
-                elif i < j:
-                    fig_fit.delaxes(axs[i, j])
-
-        for i in range(0, 5):
-            # axes.subplot(5,5,i+1+5*i)
-            axs[i, i].plot(evaluated_ranges[i, :], marginalized_1d[i, :] - np.min(sum_dev))
-            axs[i, i].plot(evaluated_ranges[i, :], np.ones(len(evaluated_ranges[i, :])) * 1., color='k')
-            axs[i, i].plot(evaluated_ranges[i, :], np.ones(len(evaluated_ranges[i, :])) * 2.7, color='k')
-            axs[i, i].yaxis.set_label_position("right")
-            axs[i, i].yaxis.tick_right()
-            axs[i, i].xaxis.set_label_position("top")
-            axs[i, i].xaxis.tick_top()
-            axs[i, i].set_xlabel(fit_values_names[i])
-
-        # axs[0,0].set_ylabel(fit_values_names[0])
-        # axs[4,4].set_xlabel(fit_values_names[4])
-        axs[4, 4].xaxis.set_label_position("bottom")
-        axs[4, 4].xaxis.tick_bottom()
-
-    # make a dictionary to return
-    fit_dict = {'fit_values': fit_values, 'fit_values_names': fit_values_names, 'sum_dev': sum_dev,
-                'fit_result': fit_result, 'marginalized_2d': marginalized_2d, 'marginalized_1d': marginalized_1d,
-                'evaluated_ranges': evaluated_ranges}  # , 'x0':x0, 'z':z}
-    return fit_dict
+    nonlinear_mag_for_plot, linear_mag_for_plot
+from submm.KIDs.res.data_io import Fit, ResSet, NonlinearIQRes, NonlinearMagRes, LinearMagRes
+from submm.KIDs.res.utils import amplitude_normalization, calc_qc_qi, guess_x0_iq_nonlinear, guess_x0_mag_nonlinear, \
+    guess_x0_iq_nonlinear_sep, guess_x0_mag_nonlinear_sep
 
 
 def fit_nonlinear_iq(f_hz, z, bounds=None, x0: list = None, fr_guess: float = None, tau=None, tau_guess=None,
                      amp_norm: bool = False, verbose: bool = True):
-    """
-    Fit a nonlinear IQ with from an S21 sweep.
+    """Fit a nonlinear IQ with from an S21 sweep.
 
     Parameters
     ----------
@@ -197,6 +68,11 @@ def fit_nonlinear_iq(f_hz, z, bounds=None, x0: list = None, fr_guess: float = No
         function of the cryostat is not flat.
     verbose : bool, optional (default True)
         Uses the print function to display fit results when true, no prints to the console when false.
+
+    Returns
+    -------
+    fit : Fit
+        A Fit NamedTuple containing the fit results.
     """
 
     if bounds is None:
@@ -258,9 +134,9 @@ def fit_nonlinear_iq(f_hz, z, bounds=None, x0: list = None, fr_guess: float = No
 
 
 def fit_nonlinear_iq_sep(fine_f_hz, fine_z, gain_f_hz, gain_z,
-                         fine_z_err=None, gain_z_err=None, bounds=None, x0=None, amp_norm: bool = False):
-    """
-    same as above function but takes fine and gain scans seporately
+                         fine_z_err=None, gain_z_err=None, bounds=None, x0=None, amp_norm: bool = False,
+                         verbose: bool = True):
+    """Same as fit_nonlinear_iq() but takes fine and gain scans separately
 
     Parameters
     ----------
@@ -296,8 +172,13 @@ def fit_nonlinear_iq_sep(fine_f_hz, fine_z, gain_f_hz, gain_z,
     amp_norm : bool, optional (default False)
         When True, a normalization is preformed for the amplitude variable. This parameter is useful when the transfer
         function of the cryostat is not flat.
+    verbose : bool, optional (default True)
+        Uses the print function to display fit results when true, no prints to the console when false.
 
-
+    Returns
+    -------
+    fit : Fit
+        A Fit NamedTuple containing the fit results.
     """
     if bounds is None:
         # define default bounds
@@ -310,48 +191,45 @@ def fit_nonlinear_iq_sep(fine_f_hz, fine_z, gain_f_hz, gain_z,
         # fr_guess = x[np.argmin(np.abs(z))]
         # x0 = [fr_guess,10000.,0.5,0,0,np.mean(np.real(z)),np.mean(np.imag(z)),3*10**-7,fr_guess]
         x0 = guess_x0_iq_nonlinear_sep(fine_f_hz, fine_z, gain_f_hz, gain_z)
-        # print(x0)
-    # User the error?
-    use_err = fine_z_err is not None and gain_z_err is not None
-
     f_hz = np.hstack((fine_f_hz, gain_f_hz))
     z = np.hstack((fine_z, gain_z))
-    if use_err:
-        z_err = np.hstack((fine_z_err, gain_z_err))
-
     if amp_norm:
         z = amplitude_normalization(f_hz, z)
-
     z_stacked = np.hstack((np.real(z), np.imag(z)))
-    if use_err:
+    # map the initial guess to the standard data record
+    guess = NonlinearIQRes(*x0)
+    if verbose:
+        guess.console(label='Guess', print_header=True)
+    if fine_z_err is not None and gain_z_err is not None:
+        z_err = np.hstack((fine_z_err, gain_z_err))
         z_err_stacked = np.hstack((np.real(z_err), np.imag(z_err)))
-        fit = optimization.curve_fit(nonlinear_iq_for_fitter, f_hz, z_stacked, x0, sigma=z_err_stacked, bounds=bounds)
-    else:
-        fit = optimization.curve_fit(nonlinear_iq_for_fitter, f_hz, z_stacked, x0, bounds=bounds)
-
-
-    # x0_result = nonlinear_iq(f_hz, x0[0], x0[1], x0[2], x0[3], x0[4], x0[5], x0[6], x0[7], x0[8])
-
-    if use_err:
-        fit_result = nonlinear_iq(f_hz, fit[0][0], fit[0][1], fit[0][2], fit[0][3], fit[0][4], fit[0][5], fit[0][6],
-                                  fit[0][7], fit[0][8])
-        # only do it for fine data
-        # red_chi_sqr = np.sum(z_stacked-np.hstack((np.real(fit_result),np.imag(fit_result))))**2/z_err_stacked**2)/(len(z_stacked)-8.)
+        popt, pcov = optimization.curve_fit(nonlinear_iq_for_fitter, f_hz, z_stacked, x0, sigma=z_err_stacked,
+                                            bounds=bounds)
+        fr, Qr, amp, phi, a, i0, q0, tau, f0 = popt
+        fit_result = nonlinear_iq(f_hz=f_hz, fr=fr, Qr=Qr, amp=amp, phi=phi, a=a, i0=i0, q0=q0, tau=tau, f0=f0)
         # only do it for fine data
         red_chi_sqr = np.sum((np.hstack((np.real(fine_z), np.imag(fine_z))) - np.hstack(
             (np.real(fit_result[0:len(fine_z)]), np.imag(fit_result[0:len(fine_z)])))) ** 2 / np.hstack(
             (np.real(fine_z_err), np.imag(fine_z_err))) ** 2) / (len(fine_z) * 2. - 8.)
-        # make a dictionary to return
+
     else:
-        # make a dictionary to return
-        fit_dict = {'fit': fit, 'fit_result': fit_result, 'x0_result': x0_result, 'x0': x0, 'z': z, 'fit_freqs': f_hz}
+        popt, pcov = optimization.curve_fit(nonlinear_iq_for_fitter, f_hz, z_stacked, x0, bounds=bounds)
+        fr, Qr, amp, phi, a, i0, q0, tau, f0 = popt
+        red_chi_sqr = None
 
-    return fit_dict
+    Qc, Qi = calc_qc_qi(qr=Qr, amp=amp)
+    result = NonlinearIQRes(fr=fr, Qr=Qr, amp=amp, phi=phi, a=a, i0=i0, q0=q0, tau=tau, f0=f0,
+                            red_chi_sqr=red_chi_sqr, Qc=Qc, Qi=Qi)
+    if verbose:
+        result.console(label='Fit', print_header=False)
+    fit = Fit(origin=inspect.currentframe().f_code.co_name, func=nonlinear_iq,
+              guess=guess, result=result, pcov=pcov, f_data=f_hz, z_data=z)
+    return fit
 
 
-# same function but double fits so that it can get error and a proper covariance matrix out
-def fit_nonlinear_iq_with_err(f_hz, z, bounds=None, x0=None, amp_norm: bool = False):
-    """
+def fit_nonlinear_iq_with_err(f_hz, z, bounds=None, x0=None, amp_norm: bool = False, verbose: bool = True):
+    """Same as fit_nonlinear_iq(), but double fits so that it can get error and a proper covariance matrix out.
+
     Parameters
     ----------
     f_hz : numpy.array
@@ -376,6 +254,13 @@ def fit_nonlinear_iq_with_err(f_hz, z, bounds=None, x0=None, amp_norm: bool = Fa
     amp_norm : bool, optional (default False)
         When True, a normalization is preformed for the amplitude variable. This parameter is useful when the transfer
         function of the cryostat is not flat.
+    verbose : bool, optional (default True)
+        Uses the print function to display fit results when true, no prints to the console when false.
+
+    Returns
+    -------
+    fit : Fit
+        A Fit NamedTuple containing the fit results.
     """
     if bounds is None:
         # define default bounds
@@ -388,25 +273,29 @@ def fit_nonlinear_iq_with_err(f_hz, z, bounds=None, x0=None, amp_norm: bool = Fa
         x0 = guess_x0_iq_nonlinear(f_hz, z)
     if amp_norm:
         z = amplitude_normalization(f_hz, z)
+    guess = NonlinearIQRes(*x0)
+    if verbose:
+        guess.console(label='Guess', print_header=True)
     z_stacked = np.hstack((np.real(z), np.imag(z)))
-    fit = optimization.curve_fit(nonlinear_iq_for_fitter, f_hz, z_stacked, x0, bounds=bounds)
-    fit_result = nonlinear_iq(f_hz, fit[0][0], fit[0][1], fit[0][2], fit[0][3], fit[0][4], fit[0][5], fit[0][6], fit[0][7],
-                              fit[0][8])
-    fit_result_stacked = nonlinear_iq_for_fitter(f_hz, fit[0][0], fit[0][1], fit[0][2], fit[0][3], fit[0][4], fit[0][5],
-                                                 fit[0][6], fit[0][7], fit[0][8])
-    x0_result = nonlinear_iq(f_hz, x0[0], x0[1], x0[2], x0[3], x0[4], x0[5], x0[6], x0[7], x0[8])
+    popt_first, pcov_first = optimization.curve_fit(nonlinear_iq_for_fitter, f_hz, z_stacked, x0, bounds=bounds)
+    fr_first, Qr_first, amp_first, phi_first, a_first, i0_first, q0_first, tau_first, f0_first = popt_first
+    
+    fit_result_stacked = nonlinear_iq_for_fitter(f_hz=f_hz, fr=fr_first, Qr=Qr_first, amp=amp_first, phi=phi_first, 
+                                                 a=a_first, i0=i0_first, q0=q0_first, tau=tau_first, f0=f0_first)
     # get error
     var = np.sum((z_stacked - fit_result_stacked) ** 2) / (z_stacked.shape[0] - 1)
     err = np.ones(z_stacked.shape[0]) * np.sqrt(var)
     # refit
-    fit = optimization.curve_fit(nonlinear_iq_for_fitter, f_hz, z_stacked, x0, err, bounds=bounds)
-    fit_result = nonlinear_iq(f_hz, fit[0][0], fit[0][1], fit[0][2], fit[0][3], fit[0][4], fit[0][5], fit[0][6], fit[0][7],
-                              fit[0][8])
-    x0_result = nonlinear_iq(f_hz, x0[0], x0[1], x0[2], x0[3], x0[4], x0[5], x0[6], x0[7], x0[8])
-
+    popt, pcov = optimization.curve_fit(nonlinear_iq_for_fitter, f_hz, z_stacked, x0, err, bounds=bounds)
+    fr, Qr, amp, phi, a, i0, q0, tau, f0 = popt
     # make a dictionary to return
-    fit_dict = {'fit': fit, 'fit_result': fit_result, 'x0_result': x0_result, 'x0': x0, 'z': z}
-    return fit_dict
+    Qc, Qi = calc_qc_qi(qr=Qr, amp=amp)
+    result = NonlinearIQRes(fr=fr, Qr=Qr, amp=amp, phi=phi, a=a, i0=i0, q0=q0, tau=tau, f0=f0, Qc=Qc, Qi=Qi)
+    if verbose:
+        result.console(label='Fit', print_header=False)
+    fit = Fit(origin=inspect.currentframe().f_code.co_name, func=nonlinear_iq,
+              guess=guess, result=result, pcov=pcov, f_data=f_hz, z_data=z)
+    return fit
 
 
 # function for fitting an iq sweep with the above equation
@@ -429,10 +318,16 @@ def fit_nonlinear_mag(f_hz, z, bounds=None, x0=None, verbose=True):
         a_guess   = x0[4]
         b0_guess  = x0[5]
         b1_guess  = x0[6]
+        flin_guess= x0[7]
         The fit's initial guess can be very important because least squares fitting does not completely search the
         parameter space.
     verbose : bool, optional (default True)
         Uses the print function to display fit results when true, no prints to the console when false.
+
+    Returns
+    -------
+    fit : Fit
+        A Fit NamedTuple containing the fit results.
     """
     if bounds is None:
         print("default bounds used")
@@ -444,29 +339,19 @@ def fit_nonlinear_mag(f_hz, z, bounds=None, x0=None, verbose=True):
         # x0 = [fr_guess,10000.,0.5,0,0,np.abs(z[0])**2,np.abs(z[0])**2,fr_guess]
         x0 = guess_x0_mag_nonlinear(f_hz, z, verbose=verbose)
 
-    fit = optimization.curve_fit(nonlinear_mag, f_hz, np.abs(z) ** 2, x0, bounds=bounds)
-    fit_result = np.sqrt(
-        nonlinear_mag(f_hz, fit[0][0], fit[0][1], fit[0][2], fit[0][3], fit[0][4], fit[0][5], fit[0][6], fit[0][7]))
-    x0_result = np.sqrt(nonlinear_mag(f_hz, x0[0], x0[1], x0[2], x0[3], x0[4], x0[5], x0[6], x0[7]))
-
+    guess = NonlinearMagRes(*x0)
     if verbose:
-        print_fit_string_nonlinear_mag(fit[0], print_header=False, label="Fit  ")
+        guess.console(label='Guess', print_header=True)
 
-    # human-readable results
-    fr = fit[0][0]
-    Qr = fit[0][1]
-    amp = fit[0][2]
-    phi = fit[0][3]
-    a = fit[0][4]
-    b0 = fit[0][5]
-    b1 = fit[0][6]
-    Qc = Qr / amp
-    Qi = 1.0 / ((1.0 / Qr) - (1.0 / Qc))
-
-    # make a dictionary to return
-    fit_dict = {'fit': fit, 'fit_result': fit_result, 'x0_result': x0_result, 'x0': x0, 'z': z,
-                'fr': fr, 'Qr': Qr, 'amp': amp, 'phi': phi, 'a': a, 'b0': b0, 'b1': b1, 'Qi': Qi, 'Qc': Qc}
-    return fit_dict
+    popt, pcov = optimization.curve_fit(nonlinear_mag, f_hz, np.abs(z) ** 2, x0, bounds=bounds)
+    fr, Qr, amp, phi, a, b0, b1, flin = popt
+    Qc, Qi = calc_qc_qi(qr=Qr, amp=amp)
+    result = NonlinearMagRes(fr=fr, Qr=Qr, amp=amp, phi=phi, a=a, b0=b0, b1=b1, flin=flin, Qc=Qc, Qi=Qi)
+    if verbose:
+        result.console(label='Fit', print_header=False)
+    fit = Fit(origin=inspect.currentframe().f_code.co_name, func=nonlinear_mag_for_plot,
+              guess=guess, result=result, pcov=pcov, f_data=f_hz, z_data=z)
+    return fit
 
 
 def fit_linear_mag(f_hz, z, bounds=None, x0=None, verbose=True):
@@ -490,6 +375,12 @@ def fit_linear_mag(f_hz, z, bounds=None, x0=None, verbose=True):
         parameter space.
     verbose : bool, optional (default True)
         Uses the print function to display fit results when true, no prints to the console when false.
+
+
+    Returns
+    -------
+    fit : Fit
+        A Fit NamedTuple containing the fit results.
     """
     if bounds is None:
         # define default bounds
@@ -497,36 +388,30 @@ def fit_linear_mag(f_hz, z, bounds=None, x0=None, verbose=True):
         bounds = ([np.min(f_hz), 100, .01, -np.pi, -np.inf], [np.max(f_hz), 200000, 1, np.pi, np.inf])
     if x0 is None:
         # define default initial guess
-        print("default initial guess used")
+        if verbose:
+            print("default initial guess used")
         # x0 = [fr_guess,10000.,0.5,0,0,np.abs(z[0])**2,np.abs(z[0])**2,fr_guess]
         x0 = guess_x0_mag_nonlinear(f_hz, z, verbose=verbose)
         x0 = np.delete(x0, [4, 6, 7])
-
-    fit = optimization.curve_fit(linear_mag, f_hz, np.abs(z) ** 2, x0, bounds=bounds)
-    fit_result = np.sqrt(linear_mag(f_hz, fit[0][0], fit[0][1], fit[0][2], fit[0][3], fit[0][4]))
-    x0_result = np.sqrt(linear_mag(f_hz, x0[0], x0[1], x0[2], x0[3], x0[4]))
-
+    guess = LinearMagRes(*x0)
     if verbose:
-        print_fit_string_linear_mag(fit[0], print_header=False, label="Fit  ")
-
+        guess.console(label='Guess', print_header=True)
+    popt, pcov = optimization.curve_fit(linear_mag, f_hz, np.abs(z) ** 2, x0, bounds=bounds)
     # human-readable results
-    fr = fit[0][0]
-    Qr = fit[0][1]
-    amp = fit[0][2]
-    phi = fit[0][3]
-    b0 = fit[0][4]
-    Qc = Qr / amp
-    Qi = 1.0 / ((1.0 / Qr) - (1.0 / Qc))
-
-    # make a dictionary to return
-    fit_dict = {'fit': fit, 'fit_result': fit_result, 'x0_result': x0_result, 'x0': x0, 'z': z,
-                'fr': fr, 'Qr': Qr, 'amp': amp, 'phi': phi, 'b0': b0, 'Qi': Qi, 'Qc': Qc}
-    return fit_dict
+    fr, Qr, amp, phi, b0 = popt
+    Qc, Qi = calc_qc_qi(qr=Qr, amp=amp)
+    result = LinearMagRes(fr=fr, Qr=Qr, amp=amp, phi=phi, b0=b0, Qc=Qc, Qi=Qi)
+    if verbose:
+        result.console(label='Fit', print_header=False)
+    fit = Fit(origin=inspect.currentframe().f_code.co_name, func=linear_mag_for_plot,
+              guess=guess, result=result, pcov=pcov, f_data=f_hz, z_data=z)
+    return fit
 
 
-def fit_nonlinear_mag_sep(fine_f_hz, fine_z, gain_f_hz, gain_z, fine_z_err=None, gain_z_err=None, bounds=None, x0=None,
-                          verbose=True):
-    """
+def fit_nonlinear_mag_sep(fine_f_hz, fine_z, gain_f_hz, gain_z, 
+                          fine_z_err=None, gain_z_err=None, bounds=None, x0=None, verbose=True):
+    """Same as, fit_nonlinear_mag(), above but fine and gain scans are provided separately.
+    
     Parameters
     ----------
     fine_f_hz : numpy.array
@@ -560,11 +445,10 @@ def fit_nonlinear_mag_sep(fine_f_hz, fine_z, gain_f_hz, gain_z, fine_z_err=None,
     verbose : bool, optional (default True)
         Uses the print function to display fit results when true, no prints to the console when false.
 
-    # same as above but fine and gain scans are provided seperatly
-    # keywords are
-    # bounds ---- which is a 2d tuple of low the high values to bound the problem by
-    # x0    --- intial guess for the fit this can be very important becuase because least square space over all the parameter is comple
-    # amp_norm --- do a normalization for variable amplitude. usefull when tranfer function of the cryostat is not flat 
+    Returns
+    -------
+    fit : Fit
+        A Fit NamedTuple containing the fit results.
     """
     if bounds is None:
         # define default bounds
@@ -572,40 +456,38 @@ def fit_nonlinear_mag_sep(fine_f_hz, fine_z, gain_f_hz, gain_z, fine_z_err=None,
         bounds = ([np.min(fine_f_hz), 100, .01, -np.pi, 0, -np.inf, -np.inf, np.min(fine_f_hz)],
                   [np.max(fine_f_hz), 1000000, 100, np.pi, 5, np.inf, np.inf, np.max(fine_f_hz)])
     if x0 is None:
-        # define default intial guess
+        # define default initial guess
         print("default initial guess used")
         x0 = guess_x0_mag_nonlinear_sep(fine_f_hz, fine_z, gain_f_hz, gain_z)
-    # use error when fitting the data?
-    use_err = fine_z_err is not None and gain_z_err is not None
-    # stack the scans for curvefit
+        
+    guess = NonlinearMagRes(*x0)
+    if verbose:
+        guess.console(label='Guess', print_header=True)
+    # stack the scans for curve_fit
     f_hz = np.hstack((fine_f_hz, gain_f_hz))
     z = np.hstack((fine_z, gain_z))
-    if use_err:
+    if fine_z_err is not None and gain_z_err is not None:
         z_err = np.hstack((fine_z_err, gain_z_err))
         # propagation of errors left out cross term
         z_err = np.sqrt(4 * np.real(z_err) ** 2 * np.real(z) ** 2 + 4 * np.imag(z_err) ** 2 * np.imag(z) ** 2)
-        fit = optimization.curve_fit(nonlinear_mag, f_hz, np.abs(z) ** 2, x0, sigma=z_err, bounds=bounds)
-    else:
-        fit = optimization.curve_fit(nonlinear_mag, f_hz, np.abs(z) ** 2, x0, bounds=bounds)
-    fit_result = nonlinear_mag(f_hz, fit[0][0], fit[0][1], fit[0][2], fit[0][3], fit[0][4], fit[0][5], fit[0][6],
-                               fit[0][7])
-    x0_result = nonlinear_mag(f_hz, x0[0], x0[1], x0[2], x0[3], x0[4], x0[5], x0[6], x0[7])
-
-    # compute reduced chi squared
-    if verbose:
-        print(f'len(z)=={len(z)}')
-    if use_err:
-        # red_chi_sqr = np.sum((np.abs(z)**2-fit_result)**2/z_err**2)/(len(z)-7.)
-        # only use fine scan for reduced chi squared.
+        popt, pcov = optimization.curve_fit(nonlinear_mag, f_hz, np.abs(z) ** 2, x0, sigma=z_err, bounds=bounds)
+        fr, Qr, amp, phi, a, b0, b1, flin = popt
+        fit_result = nonlinear_mag(f_hz=f_hz, fr=fr, Qr=Qr, amp=amp, phi=phi, a=a, b0=b0, b1=b1, flin=flin)
         red_chi_sqr = np.sum((np.abs(fine_z) ** 2 - fit_result[0:len(fine_z)]) ** 2 / z_err[0:len(fine_z)] ** 2) / (
                 len(fine_z) - 7.)
-        # make a dictionary to return
-        fit_dict = {'fit': fit, 'fit_result': fit_result, 'x0_result': x0_result, 'x0': x0, 'z': z, 'fit_freqs': f_hz,
-                    'red_chi_sqr': red_chi_sqr}
     else:
-        # make a dictionary to return
-        fit_dict = {'fit': fit, 'fit_result': fit_result, 'x0_result': x0_result, 'x0': x0, 'z': z, 'fit_freqs': f_hz}
-    return fit_dict
+        popt, pcov = optimization.curve_fit(nonlinear_mag, f_hz, np.abs(z) ** 2, x0, bounds=bounds)
+        fr, Qr, amp, phi, a, b0, b1, flin = popt
+        red_chi_sqr = None
+    # human-readable results
+    Qc, Qi = calc_qc_qi(qr=Qr, amp=amp)
+    result = NonlinearMagRes(fr=fr, Qr=Qr, amp=amp, phi=phi, a=a, b0=b0, b1=b1, flin=flin, 
+                             red_chi_sqr=red_chi_sqr, Qc=Qc, Qi=Qi)
+    if verbose:
+        result.console(label='Fit', print_header=False)
+    fit = Fit(origin=inspect.currentframe().f_code.co_name, func=nonlinear_mag_for_plot,
+              guess=guess, result=result, pcov=pcov, f_data=f_hz, z_data=z)
+    return fit
 
 
 def fit_nonlinear_iq_multi(f_hz, z, tau=None):
@@ -823,5 +705,5 @@ if __name__ == '__main__':
     res_set = ResSet(res_fits=[res_fit])
     res_set.write()
     res_set_read = ResSet(path=res_set.path)
-    for result in res_set_read:
-        result.console(label='Read', fields=['fr', 'tau', 'Qc'])
+    for read_result in res_set_read:
+        read_result.console(label='Read', fields=['fr', 'tau', 'Qc'])
